+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* [linalg] Avoid panic in Apple sysctl-based feature probing (AMX detection).
+
 # 0.22.0 - 2025-08-25
 * [Breaking][MSRV] MSRV bumped to 1.85.0
 * port to edition 2024
@@ -544,4 +546,3 @@ I wish I had seen these issues yesterday. Anyway, version numbers are cheap.
 ## Before...
 
 This Changelog started way too late. But better late than never.
-